Replace Any::Moose with plain Moose
`Any::Moose` is deprecated and thus needs to be replaced.  RT#104920
recommends replacing `Any::Moose` with `Moo` (since `Moo` provides a large
subset of the `Moose` functionality but with a shorter startup time and
lower overhead) however in order to get the enum type constraints working
properly it turned out to be much simpler to simply replace `Any::Moose`
with `Moose` (it was a simple drop-in replacement).  The tests still pass
and it doesn't seem to be any slower than before.  According to
http://shadow.cat/blog/matt-s-trout/moo-versus-any-moose/ it seems that
`Any::Moose` could still load `Moose` anyway (depending upon configuration),
so the replacement implemented here seems like a sensible solution and
resolves the issue in RT#104920.