Replace Any::Moose with Moo/Type::Tiny
RT#104920 recommends that `Any::Moose` be replaced by `Moo` since
`Any::Moose` is deprecated.  This set of changes implements this with a
combination of `Moo` and `Type::Tiny`.  It was also necessary to make
`process_tilde` no longer a required attribute in order to keep the tests
passing (`t/01-encoding.t` complained about `process_tilde` being a missing
argument); from the code it looks like `process_tilde` doesn't have to be
necessary, hence the rationale for this particular change.