Add docs to all subs and methods
A lot of the information documented here has been gleaned from the following
links:
 - http://grandzebu.net/informatique/codbar-en/datamatrix.htm
 - http://www.tcpdf.org/doc/code/classDatamatrix.html
 - http://grepcode.com/file/repo1.maven.org/maven2/com.lowagie/itext/2.0.6/com/lowagie/text/pdf/BarcodeDatamatrix.java#BarcodeDatamatrix.asciiEncodation%28byte[]%2Cint%2Cint%2Cbyte[]%2Cint%2Cint%29
 - https://github.com/itext/itextpdf/blob/master/itext/src/main/java/com/itextpdf/text/pdf/BarcodeDatamatrix.java
 - https://en.wikipedia.org/wiki/Data_Matrix
 - https://en.wikipedia.org/wiki/Reed%E2%80%93Solomon_error_correction
 - https://en.wikiversity.org/wiki/Reed%E2%80%93Solomon_codes_for_coders#Reed.E2.80.93Solomon_codes

In some of the routines (e.g. `ecc()`) it is not 100% clear exactly what is
going on, however a good attempt has been made to understand the code and/or
the context with which it is used so as to provide documentation without
being misleading, even if the documentation isn't 100% in line with the
original author's intent.