Remove unused constants
The constants `@GFI`, `@GFL` and `%POLY`, defined in
`Barcode::DataMatrix::Constants` weren't actually used.  These variables
were duplicated in `Barcode::DataMatrix::Reed`, which is where they were
used.  Thus this commit removes the unused, duplicate entries in
`Barcode::DataMatrix::Constants` and moves the documentation about the
meaning of the constants in to `Barcode::DataMatrix::Reed` so that this
information doesn't get lost.