Reformat DataMatrix code
This removes hard tabs and reformats the code according to the base perltidy
rules.  perltidy has a reasonable set of defaults for Perl code, so this
should improve the readability of the code somewhat.  By replacing the hard
tabs with spaces now the code is indented consistently in all editors, so
this is already a win.