Remove unused variable in EncodeBASE256
For some reason the code `my $k =` turns up in this routine, and happened to
be noticed while reading the code (because it looked odd).  The value `$k`
is actually assigned the value of the variable assigned on the following
line.  However, the variable `$k` is not used in the routine at all and
hence this bug didn't have an influence on the code overall and thus it can
safely be removed.